+++ legacy/bun/platform/src/actions/install_deps.rs
@@ -2,7 +2,7 @@ use moon_action::Operation;
 use moon_bun_tool::BunTool;
 use moon_console::{Checkpoint, Console};
 use moon_lang::has_vendor_installed_dependencies;
-use moon_logger::{debug, info};
+use moon_logger::{debug, error, info};
 use moon_tool::DependencyManager;
 use moon_utils::{is_ci, is_test_env};
 use std::path::Path;
@@ -29,15 +29,37 @@ pub async fn install_deps(
 
     debug!(target: LOG_TARGET, "Installing dependencies");
 
-    console
-        .out
-        .print_checkpoint(Checkpoint::Setup, "bun install")?;
+    for attempt in 1..=3 {
+        if attempt == 1 {
+            console
+                .out
+                .print_checkpoint(Checkpoint::Setup, "bun install")?;
+        } else {
+            console.out.print_checkpoint_with_comments(
+                Checkpoint::Setup,
+                "bun install",
+                [format!("attempt {attempt} of 3")],
+            )?;
+        }
 
-    operations.push(
-        Operation::task_execution("bun install")
-            .track_async(|| bun.install_dependencies(&(), working_dir, !is_test_env()))
-            .await?,
-    );
+        let mut op = Operation::task_execution("bun install");
+        let result = Operation::do_track_async(&mut op, || {
+            bun.install_dependencies(&(), working_dir, !is_test_env())
+        })
+        .await;
+
+        operations.push(op);
+
+        if let Err(error) = result {
+            if attempt == 3 {
+                return Err(error);
+            } else {
+                error!("Failed to install Bun dependencies, retrying...");
+            }
+        } else {
+            break;
+        }
+    }
 
     Ok(operations)
 }